Added peopleList to Population
Since method enumerate(Thread[] list) ThreadGroup adds to list only the active threads (the threads that are running) I added a field initialPopulationList to Population in order to keep track of all the threads (dead and alive). I also modified the consructors of Coy/Faithful...Population in order to add the Coy/Faithful... objects to the initialPopulationList.
That list will be useful to start for the first time the threads of the initial population